+++ packages/@lwc/engine-core/src/framework/api.ts
@@ -19,325 +19,28 @@ import {
     isString,
     isTrue,
     isUndefined,
-    KEY__SHADOW_RESOLVER,
     StringReplace,
     toString,
 } from '@lwc/shared';
 
-import { createText, createComment, createElement, isSyntheticShadowDefined } from '../renderer';
-import { logError, logWarn } from '../shared/logger';
+import { logError } from '../shared/logger';
 
 import { invokeEventListener } from './invoker';
 import { getVMBeingRendered } from './template';
 import { EmptyArray, EmptyObject } from './utils';
-import {
-    appendVM,
-    getAssociatedVMIfPresent,
-    getAssociatedVM,
-    removeVM,
-    rerenderVM,
-    runConnectedCallback,
-    ShadowMode,
-    SlotSet,
-    VM,
-    VMState,
-    getRenderRoot,
-    createVM,
-    hydrateVM,
-    RenderMode,
-} from './vm';
+import { isComponentConstructor } from './def';
+import { ShadowMode, SlotSet, VM, RenderMode } from './vm';
 import { LightningElementConstructor } from './base-lightning-element';
+import { VNode, VNodes, VElement, VText, VCustomElement, VComment, VElementData } from './vnodes';
 import {
-    createViewModelHook,
-    fallbackElmHook,
-    removeElmHook,
-    createChildrenHook,
-    updateNodeHook,
-    insertNodeHook,
-    removeNodeHook,
-    patchChildren,
-    patchElementPropsAndAttrs,
-    allocateChildrenHook,
+    CommentHook,
+    CustomElementHook,
+    ElementHook,
+    TextHook,
     markAsDynamicChildren,
-    hydrateChildrenHook,
-    hydrateElmHook,
-    LWCDOMMode,
 } from './hooks';
-import { getComponentInternalDef, isComponentConstructor } from './def';
-import { getUpgradableConstructor } from './upgradable-element';
-import {
-    VNode,
-    VNodes,
-    VElement,
-    VText,
-    Hooks,
-    VCustomElement,
-    VComment,
-    VElementData,
-} from './vnodes';
 
-const SVG_NAMESPACE = 'http://www.w3.org/2000/svg';
 const SymbolIterator: typeof Symbol.iterator = Symbol.iterator;
-
-const TextHook: Hooks<VText> = {
-    create: (vnode) => {
-        const { owner } = vnode;
-
-        const elm = createText(vnode.text!);
-        linkNodeToShadow(elm, owner);
-        vnode.elm = elm;
-    },
-    update: updateNodeHook,
-    insert: insertNodeHook,
-    move: insertNodeHook, // same as insert for text nodes
-    remove: removeNodeHook,
-    hydrate: (vNode: VNode, node: Node) => {
-        if (process.env.NODE_ENV !== 'production') {
-            // eslint-disable-next-line lwc-internal/no-global-node
-            if (node.nodeType !== Node.TEXT_NODE) {
-                logError('Hydration mismatch: incorrect node type received', vNode.owner);
-                assert.fail('Hydration mismatch: incorrect node type received.');
-            }
-
-            if (node.nodeValue !== vNode.text) {
-                logWarn(
-                    'Hydration mismatch: text values do not match, will recover from the difference',
-                    vNode.owner
-                );
-            }
-        }
-
-        // always set the text value to the one from the vnode.
-        node.nodeValue = vNode.text ?? null;
-        vNode.elm = node;
-    },
-};
-
-const CommentHook: Hooks<VComment> = {
-    create: (vnode) => {
-        const { owner, text } = vnode;
-
-        const elm = createComment(text);
-        linkNodeToShadow(elm, owner);
-        vnode.elm = elm;
-    },
-    update: updateNodeHook,
-    insert: insertNodeHook,
-    move: insertNodeHook, // same as insert for text nodes
-    remove: removeNodeHook,
-    hydrate: (vNode: VNode, node: Node) => {
-        if (process.env.NODE_ENV !== 'production') {
-            // eslint-disable-next-line lwc-internal/no-global-node
-            if (node.nodeType !== Node.COMMENT_NODE) {
-                logError('Hydration mismatch: incorrect node type received', vNode.owner);
-                assert.fail('Hydration mismatch: incorrect node type received.');
-            }
-
-            if (node.nodeValue !== vNode.text) {
-                logWarn(
-                    'Hydration mismatch: comment values do not match, will recover from the difference',
-                    vNode.owner
-                );
-            }
-        }
-
-        // always set the text value to the one from the vnode.
-        node.nodeValue = vNode.text ?? null;
-        vNode.elm = node;
-    },
-};
-
-// insert is called after update, which is used somewhere else (via a module)
-// to mark the vm as inserted, that means we cannot use update as the main channel
-// to rehydrate when dirty, because sometimes the element is not inserted just yet,
-// which breaks some invariants. For that reason, we have the following for any
-// Custom Element that is inserted via a template.
-const ElementHook: Hooks<VElement> = {
-    create: (vnode) => {
-        const {
-            sel,
-            owner,
-            data: { svg },
-        } = vnode;
-
-        const namespace = isTrue(svg) ? SVG_NAMESPACE : undefined;
-        const elm = createElement(sel, namespace);
-
-        linkNodeToShadow(elm, owner);
-        fallbackElmHook(elm, vnode);
-        vnode.elm = elm;
-
-        patchElementPropsAndAttrs(null, vnode);
-    },
-    update: (oldVnode, vnode) => {
-        patchElementPropsAndAttrs(oldVnode, vnode);
-        patchChildren(vnode.elm!, oldVnode.children, vnode.children);
-    },
-    insert: (vnode, parentNode, referenceNode) => {
-        insertNodeHook(vnode, parentNode, referenceNode);
-        createChildrenHook(vnode);
-    },
-    move: (vnode, parentNode, referenceNode) => {
-        insertNodeHook(vnode, parentNode, referenceNode);
-    },
-    remove: (vnode, parentNode) => {
-        removeNodeHook(vnode, parentNode);
-        removeElmHook(vnode);
-    },
-    hydrate: (vnode, node) => {
-        const elm = node as Element;
-        vnode.elm = elm;
-
-        const { context } = vnode.data;
-        const isDomManual = Boolean(
-            !isUndefined(context) &&
-                !isUndefined(context.lwc) &&
-                context.lwc.dom === LWCDOMMode.manual
-        );
-
-        if (isDomManual) {
-            // it may be that this element has lwc:inner-html, we need to diff and in case are the same,
-            // remove the innerHTML from props so it reuses the existing dom elements.
-            const { props } = vnode.data;
-            if (!isUndefined(props) && !isUndefined(props.innerHTML)) {
-                if (elm.innerHTML === props.innerHTML) {
-                    delete props.innerHTML;
-                } else {
-                    logWarn(
-                        `Mismatch hydrating element <${elm.tagName.toLowerCase()}>: innerHTML values do not match for element, will recover from the difference`,
-                        vnode.owner
-                    );
-                }
-            }
-        }
-
-        hydrateElmHook(vnode);
-
-        if (!isDomManual) {
-            hydrateChildrenHook(vnode.elm.childNodes, vnode.children, vnode.owner);
-        }
-    },
-};
-
-const CustomElementHook: Hooks<VCustomElement> = {
-    create: (vnode) => {
-        const { sel, owner } = vnode;
-        const UpgradableConstructor = getUpgradableConstructor(sel);
-        /**
-         * Note: if the upgradable constructor does not expect, or throw when we new it
-         * with a callback as the first argument, we could implement a more advanced
-         * mechanism that only passes that argument if the constructor is known to be
-         * an upgradable custom element.
-         */
-        const elm = new UpgradableConstructor((elm: HTMLElement) => {
-            // the custom element from the registry is expecting an upgrade callback
-            createViewModelHook(elm, vnode);
-        });
-
-        linkNodeToShadow(elm, owner);
-        vnode.elm = elm;
-
-        const vm = getAssociatedVMIfPresent(elm);
-        if (vm) {
-            allocateChildrenHook(vnode, vm);
-        } else if (vnode.ctor !== UpgradableConstructor) {
-            throw new TypeError(`Incorrect Component Constructor`);
-        }
-        patchElementPropsAndAttrs(null, vnode);
-    },
-    update: (oldVnode, vnode) => {
-        patchElementPropsAndAttrs(oldVnode, vnode);
-        const vm = getAssociatedVMIfPresent(vnode.elm);
-        if (vm) {
-            // in fallback mode, the allocation will always set children to
-            // empty and delegate the real allocation to the slot elements
-            allocateChildrenHook(vnode, vm);
-        }
-        // in fallback mode, the children will be always empty, so, nothing
-        // will happen, but in native, it does allocate the light dom
-        patchChildren(vnode.elm!, oldVnode.children, vnode.children);
-        if (vm) {
-            if (process.env.NODE_ENV !== 'production') {
-                assert.isTrue(
-                    isArray(vnode.children),
-                    `Invalid vnode for a custom element, it must have children defined.`
-                );
-            }
-            // this will probably update the shadowRoot, but only if the vm is in a dirty state
-            // this is important to preserve the top to bottom synchronous rendering phase.
-            rerenderVM(vm);
-        }
-    },
-    insert: (vnode, parentNode, referenceNode) => {
-        insertNodeHook(vnode, parentNode, referenceNode);
-        const vm = getAssociatedVMIfPresent(vnode.elm);
-        if (vm) {
-            if (process.env.NODE_ENV !== 'production') {
-                assert.isTrue(vm.state === VMState.created, `${vm} cannot be recycled.`);
-            }
-            runConnectedCallback(vm);
-        }
-        createChildrenHook(vnode);
-        if (vm) {
-            appendVM(vm);
-        }
-    },
-    move: (vnode, parentNode, referenceNode) => {
-        insertNodeHook(vnode, parentNode, referenceNode);
-    },
-    remove: (vnode, parentNode) => {
-        removeNodeHook(vnode, parentNode);
-        const vm = getAssociatedVMIfPresent(vnode.elm);
-        if (vm) {
-            // for custom elements we don't have to go recursively because the removeVM routine
-            // will take care of disconnecting any child VM attached to its shadow as well.
-            removeVM(vm);
-        }
-    },
-    hydrate: (vnode, elm) => {
-        // the element is created, but the vm is not
-        const { sel, mode, ctor, owner } = vnode;
-
-        const def = getComponentInternalDef(ctor);
-        createVM(elm, def, {
-            mode,
-            owner,
-            tagName: sel,
-        });
-
-        vnode.elm = elm as Element;
-
-        const vm = getAssociatedVM(elm);
-        allocateChildrenHook(vnode, vm);
-
-        hydrateElmHook(vnode);
-
-        // Insert hook section:
-        if (process.env.NODE_ENV !== 'production') {
-            assert.isTrue(vm.state === VMState.created, `${vm} cannot be recycled.`);
-        }
-        runConnectedCallback(vm);
-
-        if (vm.renderMode !== RenderMode.Light) {
-            // VM is not rendering in Light DOM, we can proceed and hydrate the slotted content.
-            // Note: for Light DOM, this is handled while hydrating the VM
-            hydrateChildrenHook(vnode.elm.childNodes, vnode.children, vm);
-        }
-
-        hydrateVM(vm);
-    },
-};
-
-function linkNodeToShadow(elm: Node, owner: VM) {
-    const { renderMode, shadowMode } = owner;
-
-    // TODO [#1164]: this should eventually be done by the polyfill directly
-    if (isSyntheticShadowDefined) {
-        if (shadowMode === ShadowMode.Synthetic || renderMode === RenderMode.Light) {
-            (elm as any)[KEY__SHADOW_RESOLVER] = getRenderRoot(owner)[KEY__SHADOW_RESOLVER];
-        }
-    }
-}
 
 function addVNodeToChildLWC(vnode: VCustomElement) {
     ArrayPush.call(getVMBeingRendered()!.velements, vnode);
